requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pdfplumber==0.11.4
python-docx==1.1.2
sentence-transformers==3.3.1
spacy
numpy==1.26.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.36
google-genai
python-dotenv==1.0.1
langdetect==1.0.9
Pillow==10.4.0
